ANimatable is implemented to Enemy class, refactoring done
