+++ src/com/codecool/snake/entities/enemies/Facebook.java
@@ -1,12 +1,11 @@
 package com.codecool.snake.entities.enemies;
 
 import com.codecool.snake.Globals;
-import com.codecool.snake.entities.Animatable;
 import com.codecool.snake.entities.Interactable;
 import com.codecool.snake.entities.snakes.SnakeHead;
 import javafx.scene.layout.Pane;
 
-public class Facebook extends Enemy implements Animatable, Interactable {
+public class Facebook extends Enemy implements Interactable {
 
 
     public Facebook(Pane pane) {
@@ -18,16 +17,6 @@ public class Facebook extends Enemy implements Animatable, Interactable {
         this.speed = 1;
         this.setPosition();
         this.setDirection();
-        //System.out.println(this.positionX + " " + this.direction);
-    }
-
-    @Override
-    public void step(){
-        if(isOutOfBounds()){
-            this.setDirection();
-        }
-        this.setX(getX() + this.heading.getX());
-        this.setY(getY() + this.heading.getY());
     }
 
     @Override
