+++ src/com/codecool/snake/entities/enemies/Spotify.java
@@ -1,12 +1,11 @@
 package com.codecool.snake.entities.enemies;
 
 import com.codecool.snake.Globals;
-import com.codecool.snake.entities.Animatable;
 import com.codecool.snake.entities.Interactable;
 import com.codecool.snake.entities.snakes.SnakeHead;
 import javafx.scene.layout.Pane;
 
-public class Spotify extends Enemy implements Animatable, Interactable {
+public class Spotify extends Enemy implements Interactable {
 
 
     public Spotify(Pane pane) {
@@ -18,16 +17,6 @@ public class Spotify extends Enemy implements Animatable, Interactable {
         this.speed = 2;
         this.setPosition();
         this.setDirection();
-        //System.out.println(this.positionX + " " + this.direction);
-    }
-
-    @Override
-    public void step(){
-        if(isOutOfBounds()){
-            this.setDirection();
-        }
-        this.setX(getX() + this.heading.getX());
-        this.setY(getY() + this.heading.getY());
     }
 
     @Override
